person-monitor/requirements.txt
paho_mqtt==1.3.1
pymongo==3.2.0
tornado==5.0.2
PyDispatcher==2.0.5